# docs/swagger.yaml
basePath: /
definitions:
...
  dto.CommonResponse:
    properties:
      data:
        $ref: '#/definitions/dto.Resp'
    type: object
...
  dto.Resp:
    properties:
      success:
        type: boolean
    type: object
...
  models.CollectionShort:
    properties:
      bannerUrl:
        type: string
      id:
        type: string
      isPrivate:
        type: boolean
      name:
        type: string
      shareString:
        type: string
      transliteration:
        type: string
      userLogin:
        type: string
    type: object
...
  models.PhotoCreate:
    properties:
      collectionId:
        type: string
      paths:
        items:
          type: string
        type: array
    type: object
  models.PhotoDataResponse:
    properties:
      data:
        $ref: '#/definitions/models.Photos'
    type: object
  models.PhotoUpdateRequest:
    properties:
      description:
        type: string
      id:
        type: string
      tags:
        items:
          type: string
        type: array
    type: object
  models.Photos:
    properties:
      author:
        $ref: '#/definitions/models.SubUsers'
      collection:
        $ref: '#/definitions/models.CollectionShort'
      collectionId:
        type: string
      commentsCount:
        type: integer
      createdAt:
        type: string
      id:
        type: integer
      likes:
        type: integer
      path:
        type: string
      tags:
        items:
          $ref: '#/definitions/models.ShortTags'
        type: array
      updatedAt:
        type: string
    type: object
  models.PhotosDataResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Photos'
        type: array
      total:
        type: integer
    type: object
...
  models.ShortTags:
    properties:
      id:
        type: string
      name:
        type: string
      primaryId:
        type: string
      seriesId:
        type: string
      slug:
        type: string
      type:
        type: string
    type: object
...
  models.SubUsers:
    properties:
      avatarUrl:
        type: string
      isPremium:
        type: boolean
      login:
        type: string
      profileName:
        type: string
    type: object
...
host: dev.lootor.me
info:
  contact: {}
  description: Документация к API Lootor.me
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Lootor
  version: "1.0"
paths:
...
  /public/photos/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.PhotoDataResponse'
      summary: получить по ид
      tags:
      - photos
  /public/photos/by_collection:
    get:
      consumes:
      - application/json
      parameters:
      - description: collectionId
        in: query
        name: collectionId
        required: true
        type: string
      - description: limit
        in: query
        name: limit
        required: true
        type: string
      - description: offset
        in: query
        name: offset
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.PhotosDataResponse'
      summary: получить по ид collection
      tags:
      - photos
  /public/photos/by_user:
    get:
      consumes:
      - application/json
      parameters:
      - description: user login
        in: query
        name: userLogin
        required: true
        type: string
      - description: limit
        in: query
        name: limit
        required: true
        type: string
      - description: offset
        in: query
        name: offset
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.PhotosDataResponse'
      summary: получить все фото по логину юзера
      tags:
      - photos
...
  /secured/feedback:
    post:
      consumes:
      - multipart/form-data
      description: Загрузка файла через multipart/form-data
      parameters:
      - description: Файл для загрузки
        in: formData
        name: file
        required: true
        type: file
      - description: Заголовок тикета
        in: query
        name: title
        required: true
        type: string
      - description: Описание тикета
        in: query
        name: description
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CommonResponse'
      summary: Создать обращение
      tags:
      - feedback
...
  /secured/photos:
    patch:
      consumes:
      - application/json
      parameters:
      - description: необходимые поля
        in: body
        name: addRequest
        required: true
        schema:
          $ref: '#/definitions/models.PhotoUpdateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.PhotoDataResponse'
      summary: update
      tags:
      - photos
    post:
      consumes:
      - application/json
      parameters:
      - description: необходимые поля
        in: body
        name: addRequest
        required: true
        schema:
          $ref: '#/definitions/models.PhotoCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.PhotosDataResponse'
      summary: Создать фотки
      tags:
      - photos
  /secured/photos/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CommonResponse'
      summary: удаление photo
      tags:
      - photos
  /secured/photos/like/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CommonResponse'
      summary: like
      tags:
      - photos
...
schemes:
- https
swagger: "2.0"